requirements.txt
httpx>=0.25.0
pwntools>=4.11.0
pyyaml>=6.0
